backend-service/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
pytest
requests
httpx
email-validator
